requirements.txt
Django>=4.2.23
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
django-environ>=0.11.2
pillow>=11.3.0